+++ anyblok_pyramid_rest_api/validator.py
@@ -163,6 +163,44 @@ def collection_post_validator(request, deserializer=None, klass=None, **kwargs):
     klass.apply_validator_schema(request, 'body', Schema, opts, base)
 
 
+def collection_patch_validator(request, deserializer=None, klass=None,
+                               **kwargs):
+    if deserializer is None:
+        deserializer = extract_cstruct
+
+    base = deserializer(request)
+    # validate the body
+    model_name = klass.get_model_name(request, base)
+    Schema = klass.get_deserialize_schema('collection_patch', model_name)
+    opts = klass.get_deserialize_opts('collection_patch')
+    klass.apply_validator_schema(request, 'body', Schema, opts, base)
+
+
+def collection_put_validator(request, deserializer=None, klass=None, **kwargs):
+    if deserializer is None:
+        deserializer = extract_cstruct
+
+    base = deserializer(request)
+    # validate the body
+    model_name = klass.get_model_name(request, base)
+    Schema = klass.get_deserialize_schema('collection_put', model_name)
+    opts = klass.get_deserialize_opts('collection_put')
+    klass.apply_validator_schema(request, 'body', Schema, opts, base)
+
+
+def collection_delete_validator(request, deserializer=None, klass=None,
+                                **kwargs):
+    if deserializer is None:
+        deserializer = extract_cstruct
+
+    base = deserializer(request)
+    # validate the body
+    model_name = klass.get_model_name(request, base)
+    Schema = klass.get_deserialize_schema('collection_delete', model_name)
+    opts = klass.get_deserialize_opts('collection_delete')
+    klass.apply_validator_schema(request, 'body', Schema, opts, base)
+
+
 def get_validator(request, deserializer=None, klass=None, **kwargs):
     if deserializer is None:
         deserializer = extract_cstruct